push results to testapi

+++ qtip/ansible_library/plugins/action/aggregate.py
@@ -32,5 +32,5 @@ class ActionModule(ActionBase):
 def aggregate(group, task_vars):
     qpi_results = [task_vars['hostvars'][host]['qpi_result'] for host in task_vars['groups'][group]]
     return {
-        'score': mean([r['score'] for r in qpi_results])
+        'score': int(mean([r['score'] for r in qpi_results]))
     }
